管理者チェックを追加

+++ src/main/java/actions/EmployeeAction.java
@@ -38,6 +38,9 @@ public class EmployeeAction extends ActionBase {
      */
     public void index() throws ServletException, IOException {
 
+        //管理者かどうかのチェック //追記
+        if (checkAdmin()) { //追記
+
         //指定されたページ数の一覧画面に表示するデータを取得
         int page = getPage();
         List<EmployeeView> employees = service.getPerPage(page);
@@ -60,8 +63,12 @@ public class EmployeeAction extends ActionBase {
         //一覧画面を表示
         forward(ForwardConst.FW_EMP_INDEX);
     }
+    }
 
         public void entryNew() throws ServletException, IOException {
+
+             //管理者かどうかのチェック //追記
+            if (checkAdmin()) { //追記
 
             putRequestScope(AttributeConst.TOKEN, getTokenId()); //CSRF対策用トークン
             putRequestScope(AttributeConst.EMPLOYEE, new EmployeeView()); //空の従業員インスタンス
@@ -69,11 +76,12 @@ public class EmployeeAction extends ActionBase {
             //新規登録画面を表示
             forward(ForwardConst.FW_EMP_NEW);
         }
+        }
 
         public void create() throws ServletException, IOException {
 
             //CSRF対策 tokenのチェック
-            if (checkToken()) {
+            if (checkAdmin() && (checkToken())) {
 
                 //パラメータの値を元に従業員情報のインスタンスを作成する
                 EmployeeView ev = new EmployeeView(
@@ -114,7 +122,10 @@ public class EmployeeAction extends ActionBase {
             }
         }
 
-                public void show() throws ServletException, IOException {
+            public void show() throws ServletException, IOException {
+
+                //管理者かどうかのチェック //追記
+                if (checkAdmin()) { //追記
 
                     //idを条件に従業員データを取得する
                     EmployeeView ev = service.findOne(toNumber(getRequestParam(AttributeConst.EMP_ID)));
@@ -131,8 +142,12 @@ public class EmployeeAction extends ActionBase {
                     //詳細画面を表示
                     forward(ForwardConst.FW_EMP_SHOW);
                 }
+            }
 
-                public void edit() throws ServletException, IOException {
+           public void edit() throws ServletException, IOException {
+
+               //管理者かどうかのチェック //追記
+               if (checkAdmin()) { //追記
 
                     //idを条件に従業員データを取得する
                     EmployeeView ev = service.findOne(toNumber(getRequestParam(AttributeConst.EMP_ID)));
@@ -150,21 +165,22 @@ public class EmployeeAction extends ActionBase {
                     //編集画面を表示する
                     forward(ForwardConst.FW_EMP_EDIT);
                 }
+           }
 
-                    public void update() throws ServletException, IOException {
+            public void update() throws ServletException, IOException {
 
-                        //CSRF対策 tokenのチェック
-                        if (checkToken()) {
-                            //パラメータの値を元に従業員情報のインスタンスを作成する
-                            EmployeeView ev = new EmployeeView(
-                                    toNumber(getRequestParam(AttributeConst.EMP_ID)),
-                                    getRequestParam(AttributeConst.EMP_CODE),
-                                    getRequestParam(AttributeConst.EMP_NAME),
-                                    getRequestParam(AttributeConst.EMP_PASS),
-                                    toNumber(getRequestParam(AttributeConst.EMP_ADMIN_FLG)),
-                                    null,
-                                    null,
-                                    AttributeConst.DEL_FLAG_FALSE.getIntegerValue());
+                    //CSRF対策 tokenのチェック
+                    if (checkAdmin() && (checkToken())) {
+                       //パラメータの値を元に従業員情報のインスタンスを作成する
+                   EmployeeView ev = new EmployeeView(
+                        toNumber(getRequestParam(AttributeConst.EMP_ID)),
+                        getRequestParam(AttributeConst.EMP_CODE),
+                        getRequestParam(AttributeConst.EMP_NAME),
+                        getRequestParam(AttributeConst.EMP_PASS),
+                        toNumber(getRequestParam(AttributeConst.EMP_ADMIN_FLG)),
+                           null,
+                           null,
+                         AttributeConst.DEL_FLAG_FALSE.getIntegerValue());
 
                             //アプリケーションスコープからpepper文字列を取得
                             String pepper = getContextScope(PropertyConst.PEPPER);
@@ -194,10 +210,10 @@ public class EmployeeAction extends ActionBase {
                     }
 
 
-                            public void destroy() throws ServletException, IOException {
+                public void destroy() throws ServletException, IOException {
 
                                 //CSRF対策 tokenのチェック
-                                if (checkToken()) {
+                                if (checkAdmin() && (checkToken())) {
 
                                     //idを条件に従業員データを論理削除する
                                     service.destroy(toNumber(getRequestParam(AttributeConst.EMP_ID)));
@@ -209,7 +225,27 @@ public class EmployeeAction extends ActionBase {
                                     redirect(ForwardConst.ACT_EMP, ForwardConst.CMD_INDEX);
                                 }
                             }
-                        }
+                private boolean checkAdmin() throws ServletException, IOException {
+
+                    //セッションからログイン中の従業員情報を取得
+                    EmployeeView ev = (EmployeeView) getSessionScope(AttributeConst.LOGIN_EMP);
+
+                    //管理者でなければエラー画面を表示
+                    if (ev.getAdminFlag() != AttributeConst.ROLE_ADMIN.getIntegerValue()) {
+
+                        forward(ForwardConst.FW_ERR_UNKNOWN);
+                        return false;
+
+                    } else {
+
+                        return true;
+                    }
+                }
+          }
+
+
+
+
